Director delegate

+++ cocos/base/CCConsole.cpp
@@ -196,7 +196,7 @@ Console::Console()
     };
 
     _maxCommands = sizeof(commands)/sizeof(commands[0]);
-	for (size_t i = 0; i < _maxCommands; ++i)
+	for (int i = 0; i < _maxCommands; ++i)
 	{
 		_commands[i] = commands[i];
 	}
